User registration

+++ base/views.py
@@ -1,4 +1,6 @@
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.db.models import Q
 from django.http import HttpResponse
@@ -13,6 +15,9 @@ from base.models import Room, Topic
 
 
 def login_page(request):
+    page = "login"
+    if request.user.is_authenticated:
+        return redirect("home")
     username = request.POST.get("username")
     password = request.POST.get("password")
 
@@ -22,17 +27,30 @@ def login_page(request):
         except:
             messages.error(request, "User does not exist")
 
-        user = authenticate(
-            request,
-            username=username,
-            password=password
-        )
+        user = authenticate(request, username=username, password=password)
 
         if user is not None:
             login(request, user)
             return redirect("home")
 
-    return render(request, "base/login_register.html", {})
+    return render(request, "base/login_register.html", {"page": page})
+
+
+def register_user(request):
+    form = UserCreationForm()
+
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "An error occurred during registration")
+
+    return render(request, "base/login_register.html", {"form": form})
 
 
 def logout_user(request):
@@ -55,6 +73,7 @@ def room_detail(request, pk: str):
     return render(request, "base/room.html", {"room": room})
 
 
+@login_required(login_url="login")
 def create_room(request):
     form = RoomForm()
     context = {"form": form}
@@ -68,9 +87,14 @@ def create_room(request):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed")
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -81,8 +105,13 @@ def update_room(request, pk):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed")
+
     if request.method == "POST":
         room.delete()
         return redirect("home")
